# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dreamers Solutions ERP APIs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of products",
                "operationId": "423f7e42ca4971f5d3d5f88cac408ce3",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductResponse"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a product",
                "operationId": "f425b83e5de0cea82c8e391e28a15886",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/products/active": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of active products",
                "operationId": "6fa8c8df31f7040d72e2d82699109ec0",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductResponse"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product by id",
                "operationId": "2fe3bea93c8144748c70d56a04ed9130",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the product",
                        "required": true,
                        "allowEmptyValue": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid product id"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update a product",
                "operationId": "9491c2cb86033310194e954a3fb2d772",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the product",
                        "required": true,
                        "allowEmptyValue": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete product",
                "operationId": "ebc8f1c81b2e9b0e47ef0930addcdf79",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the product",
                        "required": true,
                        "allowEmptyValue": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid product id"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/products/{id}/active": {
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Activate a deleted product",
                "operationId": "c17245f9c264a3b457c2c1506ee50aa0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the product",
                        "required": true,
                        "allowEmptyValue": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Invalid product id"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProductRequest": {
                "title": "ProductRequest",
                "description": "Product request body",
                "required": [
                    "productName",
                    "productTypeId",
                    "active"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier for the product",
                        "type": "integer"
                    },
                    "itemCode": {
                        "description": "Item code of the product",
                        "type": "string"
                    },
                    "productName": {
                        "description": "Name of the product",
                        "type": "string"
                    },
                    "productDescription": {
                        "description": "Description of the product",
                        "type": "string"
                    },
                    "productTypeId": {
                        "description": "Product type id",
                        "type": "integer"
                    },
                    "categoryId": {
                        "description": "Category id",
                        "type": "integer"
                    },
                    "lowStockQty": {
                        "description": "Low stock quantity threshold",
                        "type": "number"
                    },
                    "unitId": {
                        "description": "Measure unit id",
                        "type": "integer"
                    },
                    "onSale": {
                        "description": "On sale status",
                        "type": "integer"
                    },
                    "active": {
                        "description": "Active status of the product",
                        "type": "integer"
                    },
                    "companyId": {
                        "description": "Company id",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "ProductResponse": {
                "title": "ProductResponse",
                "description": "Product response model",
                "properties": {
                    "id": {
                        "description": "Unique identifier for the product",
                        "type": "integer"
                    },
                    "itemCode": {
                        "description": "Item code of the product",
                        "type": "string"
                    },
                    "productName": {
                        "description": "Name of the product",
                        "type": "string"
                    },
                    "productDescription": {
                        "description": "Description of the product",
                        "type": "string"
                    },
                    "productTypeId": {
                        "description": "Product type id",
                        "type": "integer"
                    },
                    "categoryId": {
                        "description": "Category id",
                        "type": "integer"
                    },
                    "lowStockQty": {
                        "description": "Low stock quantity threshold",
                        "type": "number"
                    },
                    "unitId": {
                        "description": "Measure unit id",
                        "type": "integer"
                    },
                    "onSale": {
                        "description": "On sale status",
                        "type": "integer"
                    },
                    "active": {
                        "description": "Active status of the product",
                        "type": "integer"
                    },
                    "companyId": {
                        "description": "Company id",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT token obtained from the login endpoint",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}